解决块内空间不足问题

+++ Decode.m
@@ -8,7 +8,7 @@ res = [];
 sum = 0;%store the sum of num in order to judge the end of the locate-map
 j = 1;%index of res
 while sum < ((blocksize-2)*(blocksize-2))/2 %locate-map
-   num = data(i)*64+data(i+1)*32+data(i+2)*16+data(i+3)*8+data(i+4)*4+data(i+5)*2+data(i+6);%calculate the number 
+   num = uint8(data(i)*64+data(i+1)*32+data(i+2)*16+data(i+3)*8+data(i+4)*4+data(i+5)*2+data(i+6));%calculate the number 
    sum = sum+num;
    res(j:j+num-1) = data(i+7);
    i = i+8;
